+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Support for Maintenance Windows History in `database`
 - Support save report in cost analysis
 - Support for Capacity Reservations
-
+- Support for Renaming Analytics Service to HeatWave
 ## 4.18.0 (March 17, 2021)
 
 ### Added